+++ OSX/libsecurity_ssl/regressions/ssl-49-sni.c
@@ -25,6 +25,7 @@
 #include <Security/SecItem.h>
 #include <Security/SecRandom.h>
 
+#include <utilities/SecCFRelease.h>
 #include <string.h>
 #include <sys/types.h>
 #include <sys/socket.h>
@@ -261,7 +262,7 @@ tests(void)
         pthread_create(&client_thread, NULL, securetransport_client_thread, client);
         pthread_create(&server_thread, NULL, securetransport_server_thread, server);
 
-        int server_err, client_err;
+        intptr_t server_err, client_err;
         pthread_join(client_thread, (void*)&client_err);
         pthread_join(server_thread, (void*)&server_err);
